+++ docs/CHANGELOG.rst
@@ -15,6 +15,8 @@ Changed
 - **BACKWARD INCOMPATIBLE:** Run data purge in a separate worker to make sure
   that listener replies to the executor within 60 seconds
 - Use batcher for spawned processes in listener
+- Increase Docker's memory limit for 100MB to make sure processes are not
+  killed when using all available memory
 
 Added
 -----